refactor: Reorganizar el context de Student. #7
